media-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
minio==7.2.0
numpy==1.24.3
scipy==1.11.4
